Make `EventNode` an interface
